backend/requirements.txt
Flask>=2.0.0
flask-cors>=3.0.0
tensorflow>=2.0.0
numpy>=1.21.0
opencv-python>=4.5.0
python-dotenv>=1.0.0
